requirements.txt
fastapi
uvicorn
datasets
langchain
openai
faiss-cpu
sentence-transformers
python-multipart
streamlit
python-dotenv
streamlit
requests

